+++ src/components/Card/CardContainer/CardContainer.jsx
@@ -1,10 +1,10 @@
-import CardBody from "../CardBody/CardBody";
 import Card from "../Card";
 import CardButtonImage from "../CardImage/CardButtonImage";
 import { useParams } from "react-router-dom";
 import { useAsync } from "../../../hooks/useAsync";
 import getRecipientMessages from "../../../apis/getRecipientMessages";
 import { useEffect, useState } from "react";
+import InfiniteScroll from "./InfiniteScroll";
 
 const CardContainer = () => {
   const LIMIT = 8;
@@ -22,16 +22,6 @@ const CardContainer = () => {
       setItems(data.results || []);
     }
   }, [data]);
-
-  const handleScroll = () => {
-    const scrollHeight = document.documentElement.scrollHeight;
-    const scrollTop = document.documentElement.scrollTop;
-    const clientHeight = document.documentElement.clientHeight;
-
-    if (fetching === false && scrollTop + clientHeight >= scrollHeight) {
-      setFetching(true);
-    }
-  };
 
   const fetchMoreData = async () => {
     const data = await getRecipientMessages({
@@ -51,28 +41,17 @@ const CardContainer = () => {
     setFetching(false);
   };
 
-  useEffect(() => {
-    if (fetching && items.length > 0) {
-      fetchMoreData();
-    }
-  }, [fetching, items]);
-
-  useEffect(() => {
-    window.addEventListener("scroll", handleScroll);
-
-    return () => {
-      window.removeEventListener("scroll", handleScroll);
-    };
-  }, []);
-
   return (
     <>
       <Card>
         <CardButtonImage id={id} />
       </Card>
-      {items?.map((item) => {
-        return <CardBody key={item.id} item={item} />;
-      })}
+      <InfiniteScroll
+        items={items}
+        status={fetching}
+        setStatus={setFetching}
+        fetchfunc={fetchMoreData}
+      />
     </>
   );
 };
